+++ source/Private/Read-FileContent.ps1
@@ -1,0 +1,6 @@
+function Read-FileContent {
+    param(
+        [string]$FilePath
+    )
+    return (Get-Content -Path $FilePath -Raw)
+}