+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added public function `Get-WebCert`.
-- Added public function `New-GraphEmailApp`.
-- Added public function `Send-GraphAppEmail`.
+- Added public function Get-WebCert.
+- Added public function New-GraphEmailApp.
+- Added public function Send-GraphAppEmail.
+- Added public function New-PatchTuesdayReport.
+- Added private function Group-UpdateByProduct.
+- Added private function Read-FileContent.
+- Added private function Show-OSUpdateSection.
 
 ### Fixed